Move Comfort Noise tests to dedicated file
